dnd: Use ClutterGesture to start DnD interactions
With the last commit we introduced ShellDndStartGesture, so now port the
Draggable to actually use that gesture, and to move over to a ClutterGrab
once the DndStartGesture has recognized.

This is shuffling a lot of things around and removing quite a bit of code.

One thing to note is that to allow users of the DndStart gesture to specify
its relationship with other gestures (eg. to allow it run together with a
LongPressGesture), the DndStartGesture is exposed on the draggable as
the startGesture property.

Also, restore the integration with WindowPreview gestures, so all of them
are able to cooperate. This used to be done via addClickAction() in the past,
we can now specify this interaction to ClutterGesture.

Part-of: <https://gitlab.gnome.org/GNOME/gnome-shell/-/merge_requests/2853>