st/scroll-bar: Port scroll bar scrolling to use ClutterGesture
Avoid the troublesome event handling here and replace it with two
ClutterPanGestures, one for the trough and one for the handle.

This is a fairly non-intrusive port, there's room for a lot of
improvements, for example making the trough react to motion events too
should be fairly easy now.

Part-of: <https://gitlab.gnome.org/GNOME/gnome-shell/-/merge_requests/2853>