accessDialog: Skip focus check for screenshot permissions
Screenshot is a special case compared to other permissions because
apps might want to hide themselves from the desktop when a screenshot
is about to be taken. In that case, this heuristic of checking if
the focus window corresponds to the application that is requesting
screenshot permissions becomes problematic.

Special case the screenshot permission to skip the focused window
check.

See also: https://github.com/flatpak/xdg-desktop-portal/pull/851