dialog: Remove incorrect disconnectObject()
It was added, as there was a previous signal disconnect in this
function, but it happened to be dead code, and the disconnectObject()
call was added in error. Remove it.

Fixes: 26235bbe5433c48b81bd3917aaf9ef14ff1929b2
Part-of: <https://gitlab.gnome.org/GNOME/gnome-shell/-/merge_requests/1349>
(cherry picked from commit 7f4db052a0b36f5409baac4f80c58550b9689d5a)